docs(samples): splits product search samples into separate files (#527)
* docs(samples): splits product search samples into separate files

* reset commented tests
