+++ samples/system-test/products.test.js
@@ -22,7 +22,7 @@ const cp = require('child_process');
 
 const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 
-const cmd = `node productSearch/products.js`;
+const cmd = `node productSearch`;
 
 const productSearch = new vision.ProductSearchClient();
 
@@ -92,7 +92,7 @@ describe(`products`, () => {
     assert.strictEqual(maybeProduct, false);
 
     let output = execSync(
-      `${cmd} createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
+      `${cmd}/createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
@@ -102,7 +102,7 @@ describe(`products`, () => {
     assert.ok(newProduct.productCategory === testProduct.productCategory);
 
     output = execSync(
-      `${cmd} deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
+      `${cmd}/deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
     );
     assert.match(output, /Product deleted./);
   });
@@ -116,13 +116,13 @@ describe(`products`, () => {
     );
     assert.strictEqual(await getProductOrFalse(newProductPath), false);
     let output = execSync(
-      `${cmd} createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
+      `${cmd}/createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
 
     output = execSync(
-      `${cmd} getProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
+      `${cmd}/getProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
@@ -136,14 +136,14 @@ describe(`products`, () => {
     assert.match(output, /Product labels:/);
 
     output = execSync(
-      `${cmd} deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
+      `${cmd}/deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
     );
     assert.match(output, /Product deleted./);
   });
 
   it(`should list products`, async () => {
     const output = execSync(
-      `${cmd} listProducts "${testProduct.projectId}" "${testProduct.location}"`
+      `${cmd}/listProducts "${testProduct.projectId}" "${testProduct.location}"`
     );
     assert.match(output, new RegExp(`Product id: ${testProduct.productId}`));
     assert.match(output, /Product labels:/);
@@ -157,12 +157,12 @@ describe(`products`, () => {
       newProductId
     );
     let output = execSync(
-      `${cmd} createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
+      `${cmd}/createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
     output = execSync(
-      `${cmd} updateProductLabels "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productKey}" "${testProduct.productValue}"`
+      `${cmd}/updateProductLabels "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productKey}" "${testProduct.productValue}"`
     );
 
     assert.match(
@@ -191,13 +191,13 @@ describe(`products`, () => {
     );
     assert.strictEqual(await getProductOrFalse(newProductPath), false);
     let output = execSync(
-      `${cmd} createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
+      `${cmd}/createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
 
     output = execSync(
-      `${cmd} deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
+      `${cmd}/deleteProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}"`
     );
     assert.match(output, /Product deleted./);
 
@@ -219,13 +219,13 @@ describe(`products`, () => {
 
     assert.strictEqual(await getProductOrFalse(newProductPath), false);
     let output = execSync(
-      `${cmd} createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
+      `${cmd}/createProduct "${testProduct.projectId}" "${testProduct.location}" "${newProductId}" "${testProduct.productDisplayName}" "${testProduct.productCategory}"`
     );
 
     assert.match(output, new RegExp(`Product name: ${newProductPath}`));
 
     output = execSync(
-      `${cmd} purgeOrphanProducts "${testProduct.projectId}" "${testProduct.location}"`
+      `${cmd}/purgeOrphanProducts "${testProduct.projectId}" "${testProduct.location}"`
     );
 
     assert.match(output, new RegExp(`Orphan products deleted.`));
